安装向导后会弹框提示是否继续，这个是为了明确向导已经结束，后续将无人职守安装。
但是向导已经完成，重复安装时向导不需要重复确认，此时需要等待向导流程结束弹框确认后才无人职守安装。

因此提供参数可以跳过弹框，输入密码后就无人职守安装。

+++ src/manager/flow.sh
@@ -85,10 +85,12 @@ function manager::flow::install::install() {
     # 运行安装指引
     manager::flow::apps::do_command "install_guide" || return "$SHELL_FALSE"
 
-    tui::confirm "Install Wizard completed. Next Step?"
-    if [ "$?" -ne "$SHELL_TRUE" ]; then
-        lwarn "install wizard completed, user canceled continue."
-        return "$SHELL_FALSE"
+    if manager::flags::continue_after_guide::is_not_exists; then
+        tui::confirm "Install Wizard completed. Next Step?"
+        if [ "$?" -ne "$SHELL_TRUE" ]; then
+            lwarn "install wizard completed, user canceled continue."
+            return "$SHELL_FALSE"
+        fi
     fi
 
     # 运行安装
